requirements.txt
torch
numpy
pandas
flask
fastapi
uvicorn
requests
websockets
python-dotenv
